Change polling give up policy for OneView client
On previous versions of the OneView client, the polling of changes
on OneView was done until a certain ammount of tries then stop.
Configuring such ammount of retries is very troublesome since we
need to experiment until find an acceptable ammount of retries.
To ease it, we're changing the retry policy to consider a timeout
variable in seconds. This way, it'll be easy to operators to define
a acceptable service level requirement.

Change-Id: I392e0b17e160fbc37d19183f5890038818e6c4f4
Partial-bug: #1531918