Change polling give up policy for OneView client
On previous versions of the OneView client, the polling of changes
on OneView was done until a certain amount of tries then stopped
even for tasks that are ongoing. This patch changes that behavior
to keep checking tasks until completion and use the
max_polling_attempts parameter only for exceptional cases such as
request errors/timeouts.

Change-Id: I392e0b17e160fbc37d19183f5890038818e6c4f4
Closes-Bug: #1531918