Adding fancy exceptions messages
OneViewException and its subclasses are meant to have a default message that can
be overwritten by passing a custom message in the instantiation method. This
patch ensure it happens without aid of the builtin class Exception and
overwrite __str__ to include information about the class of the exception being
thrown to ease debug.

To obtain just the exception message, users can still access the exc.message
variable.

Change-Id: I278e38dadf9b612009156ecfce62a0d31e9caa94