Reverts removal of from_json on ServerProfile cls
The method was doing something different from the one on OneViewObject: It was
adding keys on the ServerProfile object that weren't on the attribute_map.
Without it, Bad Requests are being sent to OneView since they doesn't had the
required parameters.

Reverting it back and fixing the problem of the instatiation on the old method.

Change-Id: I4e52209153a0184df274cade6db553b891bf0da6