Fixing __getattribute__ def and refactor from_json
The previous commit left a __getattribute__ definition in models.OneViewObject
that would enter an infinite loop when called with an argument that isn't on the
attribute_map of the instance. I'm removing it since it was legacy of a test and
shoudn't have remained on the code. This also refactors duplicate code on
*_from_json methods.

Change-Id: Ib080d58d666e39865cc7c03121d0341361c9617c