Updating requirements by hand
In recent run of devstack, I found that python-oneviewclient was installing a
version of pbr that wasn't the most up to date version used by openstack
services. In consequence of that, services restarted after the build of devstack
were failing. To solve that, a couple of commits were proposed on
openstack-infra/project-config and openstack/requirements so as the "Openstack
Proposal Bot" proposes changes in python-oneviewclient when the
global-requirements changes. But until the patches are approved,
I'm proposing this patch on emergency basis to update the requirements file and
keep the integration Ironic/OneView working.

Change-Id: I2a034117ad902774a613bc3c8088ac72413242e9