Bugfix+removing invalid tests+functional at gate
A bugfix was introduced on the code along with ClientV2.
`get_server_profile_template` was calling the manager of server_profile instead
of server_profile_template. This wasn't seen because functional tests for
Client(v1) were wrong and useless. Tests fixed for the case above and I changed
.testr.conf to run funcional tests on py34 and py27 gates since they're mocked
at the limits of the system anyway.

Change-Id: Iea97095dbf34267a053dafbee0ebcd56ec55db13