Add auditing instrumentation for python-oneviewclient
This change is about adding the ability to python-oneviewclient
registering time, parameters values and return values for methods
calls that does requests to OneView appliance. The instrumentation
implemented by this patch provides the necessary support that will
be needed to measure the performance of the OneView driver for
ironic and other internal tools.

Change-Id: Iae7fe8ec559ccb2cbf8f93486eb4a1a674d7053a
Co-Authored-By: Hugo Nicodemos <nicodemos@lsd.ufcg.edu.br>