Passing SUPPORTED_ONEVIEW_VERSION as string + transient error fix
requests==2.11.0 is now checking header parameters for string or buffer types.
We were passing SUPPORTED_ONEVIEW_VERSION as int, but now it's not allowed.
This patch str's SUPPORTED_ONEVIEW_VERSION before passing it to requests.

This patch also fixes a transient error occuring due to simultaneous access to
constants in 'fixtures.py'.

Change-Id: I188914c0f1c6a1c8496da59857e99ff03a2b83e6