Ensure values under validation have same data type
This patch address the case of validating resources like memory_mb
and cpus that have same values but doesn't have same type, for
example, in case of node's memory_mb and cpus registered as unicode
after an inspection process. In this patch we convert the data into
to string, making possible the values pass in the validation when
correct.

Change-Id: I014ace14cb315444643c7e2c8d5b43b01cc7ef86